requirements.txt
torch
transformers
torchvision
Pillow
streamlit
